index all can now be load balanced #1757

+++ src/main/java/edu/harvard/iq/dataverse/search/IndexAllServiceBean.java
@@ -16,6 +16,9 @@ import javax.ejb.Asynchronous;
 import javax.ejb.EJB;
 import javax.ejb.Stateless;
 import javax.inject.Named;
+import javax.json.Json;
+import javax.json.JsonArrayBuilder;
+import javax.json.JsonObjectBuilder;
 import org.apache.solr.client.solrj.SolrServer;
 import org.apache.solr.client.solrj.SolrServerException;
 import org.apache.solr.client.solrj.impl.HttpSolrServer;
@@ -38,30 +41,68 @@ public class IndexAllServiceBean {
     SystemConfig systemConfig;
 
     @Asynchronous
-    public Future<String> indexAll() {
+    public Future<JsonObjectBuilder> indexAllOrSubset(long numPartitions, long partitionId, boolean previewOnly) {
+        JsonObjectBuilder response = Json.createObjectBuilder();
+        Future<String> responseFromIndexAllOrSubset = indexAllOrSubset(numPartitions, partitionId);
+        String status = "indexAllOrSubset has begun";
+        response.add("responseFromIndexAllOrSubset", status);
+        return new AsyncResult<>(response);
+    }
+
+    public JsonObjectBuilder indexAllOrSubsetPreview(long numPartitions, long partitionId) {
+        JsonObjectBuilder response = Json.createObjectBuilder();
+        JsonObjectBuilder previewOfWorkload = Json.createObjectBuilder();
+        JsonObjectBuilder dvContainerIds = Json.createObjectBuilder();
+        JsonArrayBuilder dataverseIds = Json.createArrayBuilder();
+        List<Dataverse> dataverses = dataverseService.findAllOrSubset(numPartitions, partitionId);
+        for (Dataverse dataverse : dataverses) {
+            dataverseIds.add(dataverse.getId());
+        }
+        JsonArrayBuilder datasetIds = Json.createArrayBuilder();
+        List<Dataset> datasets = datasetService.findAllOrSubset(numPartitions, partitionId);
+        for (Dataset dataset : datasets) {
+            datasetIds.add(dataset.getId());
+        }
+        dvContainerIds.add("dataverses", dataverseIds);
+        dvContainerIds.add("datasets", datasetIds);
+        previewOfWorkload.add("dvContainerIds", dvContainerIds);
+        previewOfWorkload.add("dataverseCount", dataverses.size());
+        previewOfWorkload.add("datasetCount", datasets.size());
+        previewOfWorkload.add("partitionId", partitionId);
+        response.add("previewOfPartitionWorkload", previewOfWorkload);
+        return response;
+    }
+
+    public Future<String> indexAllOrSubset(long numPartitions, long partitionId) {
         long indexAllTimeBegin = System.currentTimeMillis();
         String status;
-        SolrServer server = new HttpSolrServer("http://" + systemConfig.getSolrHostColonPort() + "/solr");
-        logger.info("attempting to delete all Solr documents before a complete re-index");
-        try {
-            server.deleteByQuery("*:*");// CAUTION: deletes everything!
-        } catch (SolrServerException | IOException ex) {
-            status = ex.toString();
-            logger.info(status);
-            return new AsyncResult<>(status);
-        }
-        try {
-            server.commit();
-        } catch (SolrServerException | IOException ex) {
-            status = ex.toString();
-            logger.info(status);
-            return new AsyncResult<>(status);
+
+        String resultOfClearingIndexTimes;
+        if (numPartitions == 1) {
+            SolrServer server = new HttpSolrServer("http://" + systemConfig.getSolrHostColonPort() + "/solr");
+            logger.info("attempting to delete all Solr documents before a complete re-index");
+            try {
+                server.deleteByQuery("*:*");// CAUTION: deletes everything!
+            } catch (SolrServerException | IOException ex) {
+                status = ex.toString();
+                logger.info(status);
+                return new AsyncResult<>(status);
+            }
+            try {
+                server.commit();
+            } catch (SolrServerException | IOException ex) {
+                status = ex.toString();
+                logger.info(status);
+                return new AsyncResult<>(status);
+            }
+
+            int numRowsAffected = dvObjectService.clearAllIndexTimes();
+            resultOfClearingIndexTimes = "Number of rows affected by clearAllIndexTimes: " + numRowsAffected + ".";
+        } else {
+            resultOfClearingIndexTimes = "Solr index was not cleared before indexing.";
         }
 
-        int numRowsAffected = dvObjectService.clearAllIndexTimes();
-        String resultOfClearingIndexTimes = "Number of rows affected by clearAllIndexTimes: " + numRowsAffected + ".";
-
-        List<Dataverse> dataverses = dataverseService.findAll();
+        List<Dataverse> dataverses = dataverseService.findAllOrSubset(numPartitions, partitionId);
         int dataverseIndexCount = 0;
         for (Dataverse dataverse : dataverses) {
             dataverseIndexCount++;
@@ -70,7 +111,7 @@ public class IndexAllServiceBean {
         }
 
         int datasetIndexCount = 0;
-        List<Dataset> datasets = datasetService.findAll();
+        List<Dataset> datasets = datasetService.findAllOrSubset(numPartitions, partitionId);
         for (Dataset dataset : datasets) {
             datasetIndexCount++;
             logger.info("indexing dataset " + datasetIndexCount + " of " + datasets.size());
